+++ main.go
@@ -3,32 +3,46 @@ package main
 import (
 	"flag"
 	"fmt"
+	"log"
 	"os"
+	"path/filepath"
+	"sync"
 )
 
 var (
-	imgPath = flag.String("imgPath", "./DarkBB8.png", "Path to an image")
+	imgPath = flag.String("imgPath", "./imgs", "Path to a folder")
+	wg      sync.WaitGroup
 )
 
 func main() {
 	flag.Parse()
 
-	// Open the img file for decoding
-	f, err := os.Open(*imgPath)
-	defer f.Close()
+	err := filepath.Walk(*imgPath, visit)
 	if err != nil {
-		panic(err)
+		log.Println(err)
 	}
 
-	// Create the file to be written
-	newF, err := os.Create("compressed.png")
-	defer newF.Close()
-	if err != nil {
-		panic(err)
+	// Wait for all compression routines to finish before exiting
+	wg.Wait()
+}
+
+// Each image that is "visited" will be encoded/compressed in an attempt to save storage
+func visit(path string, f os.FileInfo, err error) error {
+	if f.IsDir() {
+		// Exit the function if a directory is found
+		return nil
 	}
 
-	// Call the compress func (program will hang)
-	compress(f, newF, High)
+	fmt.Printf("\nVisited: %s\n", path)
+	file, err := os.Open(path)
+	if err != nil {
+		return err
+	}
 
-	fmt.Println("Finished encoding image")
+	wg.Add(1)
+	go compress(file, High)
+
+	fmt.Printf("Finished encoding image: %s\n", f.Name())
+
+	return nil
 }