+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # CHANGELOG
+
+## Release v0.0.1
+### Added
+-   Directory tree walking
+-   Faster image compression
+-   Asynchronous file encoding
+-   Better file creation
+-   Better errors/logging to stdout